+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## Unreleased
+## 3.0.5 - 2025-02-06
 
 - **BC Break!**: Added `$page` and `$aliasNumberPages` parameters to
   the `header()` and `footer()` functions.